requirements.txt
numpy==2.3.3
pandas==2.3.3
scikit-learn==1.7.2
matplotlib==3.10.7
PyYAML==6.0.3
